GotMazeSystemFixed
I divided maze generator into constructor and designer.
It is no longer generator.

I will change some singleton pattern better way.

Right now the problem is that I cannot start the game because the game controller object isn't in the main scene.
And in the main scene game controller is not needed.